Create file for fake-promoter heuristic
Unify all related code into it.
- Remove modifi_db.py file.
- Remove helper procedures' definitions from bot_detector.py file
